+++ lambda.js
@@ -1,0 +1,4 @@
+const serverless = require('serverless-http');
+const app = require('./app');
+
+module.exports.handler = serverless(app);
