+++ package.json
@@ -3,9 +3,14 @@
   "version": "1.0.0",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js"
+    "start": "node server.js",
+    "deploy": "npx serverless deploy"
   },
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "serverless-http": "^3.0.0"
+  },
+  "devDependencies": {
+    "serverless": "^3.38.0"
   }
 }
