+++ serverless.yml
@@ -1,0 +1,11 @@
+service: yasukaribike-server
+frameworkVersion: '3'
+provider:
+  name: aws
+  runtime: nodejs18.x
+  region: ap-northeast-1
+functions:
+  app:
+    handler: lambda.handler
+    events:
+      - httpApi: '*'
